Move notification code out of `models` module
It's no longer related to the mongodb-backed models
